feat: Filter closed accounts in ClientsDataSource
This commit adds a filter to the ClientsDataSource class to exclude closed accounts from the list of clients. The filter is applied when the `showClosedAccounts` flag is set to false. This ensures that only active accounts are displayed to the user.

The code changes in this commit modify the `getClientsByCountry` and `getClients` methods to apply the filter. The `clients.pageItems` array is filtered to remove any clients with a status of "Closed".